+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 ### Added
 - Added [`no-proxy`] rule
 - Added [`no-mutating-methods`] rule
+- Added error reports when using `object.__defineSetter__()` or `object.__defineGetter__()` in [`no-get-set`] rule
 ### Fixed
 - Fixed error in [`no-mutating-assign`] report message